+++ client/index.html
@@ -5,7 +5,7 @@
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="stylesheet" href="\src\styles\index.scss" />
-    <title>Vite App</title>
+    <title>TO-DO App</title>
   </head>
   <body>
     <div id="app">
@@ -18,6 +18,6 @@
         <button id="button">Add Task</button>
       </footer>
     </div>
-    <script type="module" src="./src/lib/index.js"></script>
+    <script type="module" src="./src/main.js"></script>
   </body>
 </html>
